Don't mutate DOM in compatibility utility
