+++ views/mdc/assets/js/app.js
@@ -1,16 +1,23 @@
 import isCompatible from './utils/compatibility';
+import config from './config';
 
 document.addEventListener('DOMContentLoaded', () => {
     if (!isCompatible) {
+        const errorMessage = config.get(
+            'compatibility.errorMessage',
+            'Unsupported browser!',
+        );
+        document.body.innerHTML = `<p>${errorMessage}</p>`;
+
         return;
     }
 
-    const initialize = require('./components/initialize').initialize;
+    document.documentElement.classList.remove('incompatible-browser');
 
     window.dialogPolyfill = require('./dialog-polyfill');
     require('material-design-lite/material');
     require('./mdl-stepper');
     require('babel-polyfill');
 
-    initialize();
+    require('./components/initialize').initialize();
 });
